seguimos

+++ functions.hpp
@@ -14,5 +14,5 @@ void equalizationImageSlides(cv::Mat & image, std::string newImage, int r);
 void RGB(cv::Mat image, std::string newImage);
 void biequalization(cv::Mat image, std::vector<int> & normalizado, std::string newImage);
 void biequalizationImage(cv::Mat & image, std::string newImage);
-double calcularMediana(std::vector<int> a);
+int calcularMediana(std::vector<int> a);
 #endif
